+++ R/align_pcdirs.R
@@ -8,6 +8,8 @@
 #'
 #' @param train_obj Output object from `prep_training_data` (jfpca only)
 #'
+#' @export align_pcdirs
+#'
 #' @return List with the same structure as `prep_training_data`, but
 #' the principal directions are replaced with the aligned version and gamI is
 #' included in the fpca_res object.
@@ -23,9 +25,6 @@ align_pcdirs <- function(train_obj) {
   # Extract and prepare objects based on training data
   aligned = train_obj$alignment
   gam = t(aligned$warping_functions)
-  N = nrow(gam)
-  M = ncol(gam)
-  mq = aligned$mqn
   pc_dirs = train_obj$fpca_res$f_pca
 
   # Compute gamma inverse